Clean out miscellaneous references to componentType
Issue #1956
