--- src/app/components/body-units/aligned-image.html
@@ -1,6 +1,0 @@
-<template args="model">
-    <figure class="{model.alignmentClass}">
-        <img src="{model.imageUrl}" alt="{model.imageAlt}">
-        <figcaption data-html="caption"></figcaption>
-    </figure>
-</template>
