--- src/app/components/modal-content/modal-content.scss
@@ -1,15 +1,0 @@
-@import 'pattern-library/core/pattern-library/headers';
-
-modal-content {
-    align-items: center;
-    background-color: rgba(ui-color(black), 0.8);
-    display: flex;
-    height: 100vh;
-    justify-content: center;
-    left: 0;
-    padding: 2rem;
-    position: fixed;
-    top: 0;
-    width: 100%;
-    z-index: 5000;
-}
