--- src/app/components/sticky-footer/sticky-footer.html
@@ -1,11 +1,0 @@
-<template args="model">
-    <div class="button-group">
-        <a href="{model.leftButton.link}" class="btn medium">{model.leftButton.text}</a>
-        <div if="model.leftButton.description" class="description">{model.leftButton.description}</div>
-        <div if="model.leftButton.descriptionHtml" class="description" data-html="leftButton.descriptionHtml"></div>
-    </div>
-    <div if="model.rightButton" class="button-group">
-        <div class="description">{model.rightButton.description}</div>
-        <a href="{model.rightButton.link}" class="btn medium">{model.rightButton.text}</a>
-    </div>
-</template>
