--- src/app/pages/errata-form/banned-notice/banned-notice.html
@@ -1,3 +1,0 @@
-<template args="model">
-    {model.text}
-</template>
