--- src/app/pages/errata-form/banned-notice/banned-notice.scss
@@ -1,5 +1,0 @@
-@import 'pattern-library/core/pattern-library/headers';
-
-.banned-notice {
-    padding: $normal-margin;
-}
