--- src/app/pages/hero-journey/hero-journey.scss
@@ -1,85 +1,0 @@
-@import 'pattern-library/core/pattern-library/headers';
-
-$near-white: #fafafa;
-
-.hero-journey {
-    background-color: $near-white;
-
-    section {
-        overflow: hidden;
-        padding: 5rem 0;
-        position: relative;
-    }
-
-    .boxed {
-        display: grid;
-        grid-gap: 4rem;
-        text-align: left;
-
-        @include wider-than($phone-max) {
-            padding-right: 7.5rem; // navigator gutter
-        }
-
-        > * {
-            margin: 0;
-        }
-    }
-
-    .card {
-        @extend %card;
-
-        background-color: ui-color(white);
-        color: text-color(normal);
-        display: grid;
-        grid-gap: 2rem;
-        padding: 2rem;
-    }
-
-    .btn.primary {
-        @extend %button;
-        @extend %primary;
-
-        justify-self: left;
-    }
-
-    .pulsing {
-        align-self: start;
-        justify-self: end;
-        position: relative;
-
-        > * {
-            grid-column-start: 1;
-            grid-row-start: 1;
-            position: relative;
-        }
-
-        .pulse {
-            animation: pulse 3s ease-out;
-            animation-iteration-count: infinite;
-            background-color: os-color(orange);
-            border-radius: 0.4rem;
-            bottom: 0;
-            left: 0;
-            position: absolute;
-            right: 0;
-            top: 0;
-        }
-    }
-
-    @keyframes pulse {
-        33% {
-            opacity: 0.1;
-            transform: scale(1);
-        }
-
-        66% {
-            opacity: 0.7;
-            transform: scale(1.1, 1.3);
-        }
-
-        100% {
-            opacity: 0;
-            transform: scale(1.2, 1.5);
-        }
-    }
-}
