--- src/app/pages/hero-journey/numbered-navigator/numbered-navigator.html
@@ -1,18 +1,0 @@
-<template args="model">
-    <if condition="model">
-        <div class="right-tooltip {model.hiddenClass}">{model.tooltipMessage}</div>
-        <div class="left-tooltip {model.hiddenClass}">{model.tooltipMessage}</div>
-        <div class="phone-tooltip {model.hiddenClass}">{model.tooltipMessage}</div>
-        <div each="node in model.steps" class="node-and-bar {model.nodeStatus($item)}">
-            <div if="$item > 0" class="between-bar"></div>
-            <div class="node">
-                <if condition="model.nodeStatus($item) === 'done'">
-                    <span class="fa fa-check"></span>
-                <else>
-                    {$item + 1}
-                </if>
-            </div>
-        </div>
-        <img class="badge {model.nodeStatus(4) === 'done' ? '' : 'not-yet'}" src="/images/hero-journey/hero-badge.svg">
-    </if>
-</template>
