--- src/app/pages/hero-journey/numbered-navigator/numbered-navigator.scss
@@ -1,232 +1,0 @@
-@import 'pattern-library/core/pattern-library/headers';
-@import 'mixins/triangle';
-
-$node-size: 3.2rem;
-$shadow: 0 0.1rem 0.3rem 0 rgba(0, 0, 0, 0.2), 0 0.2rem 0.2rem 0 rgba(0, 0, 0, 0.12), 0 0 0.2rem 0 rgba(0, 0, 0, 0.14);
-$tooltip-breakpoint: #{$media-content-max + 10em};
-
-.numbered-navigator {
-    background-color: ui-color(white);
-    box-shadow: $shadow;
-    display: grid;
-    padding: 0.9rem;
-    position: fixed;
-    z-index: 1;
-
-    @include width-up-to($phone-max) {
-        grid-auto-flow: column;
-        left: 0;
-        padding-right: 7rem;
-        top: 5.5rem;
-        width: 100%;
-    }
-
-    @include wider-than($phone-max) {
-        border-radius: 2.5rem;
-        top: 12.5rem;
-        width: 5rem;
-    }
-
-    @include width-between($phone-max, $media-content-max) {
-        right: 1rem;
-    }
-
-    @include wider-than($media-content-max) {
-        right: calc(50vw - #{$media-content-max / 2});
-    }
-
-    .left-tooltip,
-    .right-tooltip,
-    .phone-tooltip {
-        align-items: center;
-        background-color: ui-color(white);
-        border-radius: 0.2rem;
-        display: grid;
-        filter: drop-shadow(rgba(0, 0, 0, 0.2) 0 0.2rem 0.2rem);
-        font-weight: bold;
-        line-height: normal;
-        min-height: 7rem;
-        padding: 1rem;
-        position: absolute;
-        // Engage 3d optimizations to eliminate artifacts in Safari
-        transform: translate3d(0, 0, 0);
-        transition: top 1s, left 1s, opacity 1s;
-        width: 11rem;
-
-        &.invisible {
-            opacity: 0;
-        }
-    }
-
-    .right-tooltip {
-        left: calc(100% + 1rem);
-
-        @include width-up-to($tooltip-breakpoint) {
-            display: none;
-        }
-
-        &::before {
-            @include triangle(
-                $direction: left,
-                $position: top calc(50% - 0.7rem) left -0.7rem,
-                $color: ui-color(white),
-                $size: 0.7rem
-            );
-        }
-    }
-
-    .left-tooltip {
-        display: none;
-        right: calc(100% + 1rem);
-
-        @include width-between($phone-max, $tooltip-breakpoint) {
-            display: inherit;
-        }
-
-        &::after {
-            @include triangle(
-                $direction: right,
-                $position: top calc(50% - 0.7rem) right -0.7rem,
-                $color: ui-color(white),
-                $size: 0.7rem
-            );
-        }
-    }
-
-    .phone-tooltip {
-        display: none;
-        top: calc(100% + 1rem);
-
-        @include width-up-to($phone-max) {
-            display: inherit;
-        }
-
-        &::before {
-            @include triangle(
-                $direction: top,
-                $position: left calc(50% - 0.7rem) top -0.3rem,
-                $color: ui-color(white),
-                $size: 0.7rem
-            );
-        }
-    }
-
-    .node-and-bar {
-        display: grid;
-
-        @include width-up-to($phone-max) {
-            align-items: center;
-            grid-auto-flow: column;
-        }
-
-        @include wider-than($phone-max) {
-            grid-template-rows: auto 1fr;
-            justify-items: center;
-        }
-
-        &:not(.done):not(.on) {
-            .node {
-                background-position: right bottom;
-                border-color: ui-color(form-border);
-                color: ui-color(form-border);
-            }
-
-            .between-bar {
-                background-color: ui-color(form-border);
-            }
-        }
-
-        &.done {
-            .node {
-                //background-color: os-color(green);
-                border-color: os-color(green);
-                color: text-color(white);
-
-                @include width-up-to($phone-max) {
-                    background-position: left bottom;
-                }
-
-                @include wider-than($phone-max) {
-                    background-position: right top;
-                }
-            }
-
-            .between-bar {
-                background-color: os-color(green);
-            }
-        }
-
-        &.on {
-            .node {
-                background-position: right bottom;
-                border-color: os-color(green);
-                color: os-color(green);
-            }
-
-            .between-bar {
-                background-color: os-color(green);
-            }
-        }
-    }
-
-    .node {
-        align-content: center;
-        border-radius: 50%;
-        border-style: solid;
-        border-width: 0.15rem;
-        cursor: pointer;
-        display: grid;
-        height: $node-size;
-        justify-content: center;
-        transition: color 0.5s linear, background-color 1s linear, border-color 1s linear, background-position 0.7s linear;
-        width: $node-size;
-
-        @include width-up-to($phone-max) {
-            background-image: linear-gradient(to right, os-color(green) 50%, transparent 50%);
-            background-size: 205% 100%;
-        }
-
-        @include wider-than($phone-max) {
-            background-image: linear-gradient(to bottom, os-color(green) 50%, transparent 50%);
-            background-size: 100% 205%;
-        }
-    }
-
-    .between-bar {
-        transition: background-color 1s linear;
-
-        @include width-up-to($phone-max) {
-            height: 0.2rem;
-            width: calc(25vw - 6rem); // Correct for 5 nodes
-        }
-
-        @include wider-than($phone-max) {
-            height: 5rem;
-            width: 0.2rem;
-        }
-    }
-
-    .badge {
-        position: absolute;
-        width: 8rem;
-
-        @include width-up-to($phone-max) {
-            right: 0;
-            top: 50%;
-            transform: translateY(-50%);
-            width: 6.5rem;
-        }
-
-        @include wider-than($phone-max) {
-            bottom: 0;
-            right: 50%;
-            transform: translate(52%, 100%);
-        }
-
-        &.not-yet {
-            filter: grayscale(100%);
-            opacity: 0.9;
-        }
-    }
-
-}
