--- src/app/pages/hero-journey/sections/quiz/quiz.html
@@ -1,34 +1,0 @@
-<template args="model">
-    <div class="boxed">
-        <div class="headline-group">
-            <h2>{model.heading}</h2>
-            <div class="skip-link">
-                Or, <a href="{model.skipLink.link}">{model.skipLink.text}</a>.
-            </div>
-        </div>
-        <div class="quiz-card card">
-            <div class="content-row">
-                <div class="card-content" each="qa in model.questions">
-                    <div class="question-number">{1 + $item}/{model.questions.length}</div>
-                    <div class="question">{qa.question}</div>
-                    <div class="answers">
-                        <div each="answer in qa.answers" class="answer card">
-                            <div class="card-content">
-                                <span class="longest-answer">{model.longestAnswer}</span>
-                                <span class="displayed-answer">{answer}</span>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="final card-content">
-                    {model.completeMessage}
-                </div>
-            </div>
-            <img if="model.currentQuestion >= model.questions.length"
-                class="big-checkmark" src="/images/hero-journey/2-quizb.svg"
-                alt="big green checkmark">
-            <img class="bottom-border" src="/images/hero-journey/card-accent.svg">
-        </div>
-    </div>
-    <img class="lower-left" src="{model.image.image}" alt="{model.image.altText}" role="decoration">
-</template>
