--- src/app/pages/hero-journey/sections/quiz/quiz.scss
@@ -1,186 +1,0 @@
-@import 'pattern-library/core/pattern-library/headers';
-
-$guessed-color: #c1c1c1;
-
-@mixin normal-padding() {
-    @include width-up-to($phone-max) {
-        padding: 2rem;
-    }
-
-    @include width-between($phone-max, $tablet-max) {
-        padding: 3rem;
-    }
-
-    @include wider-than($tablet-max) {
-        padding: 5rem;
-    }
-}
-
-.hero-journey .quiz {
-    background-image: linear-gradient(to bottom, os-color(gold) 30%, transparent 30%);
-    padding: 5rem 0;
-
-    .headline-group {
-        display: grid;
-        grid-gap: 1rem;
-
-        > * {
-            margin: 0;
-        }
-
-        h2 {
-            @include set-font(h1);
-        }
-    }
-
-    .skip-link {
-        @include set-font(h3);
-
-        a {
-            display: inline;
-        }
-    }
-
-    .question-number,
-    .question {
-        @include width-up-to($phone-max) {
-            @include set-font(h3);
-
-            line-height: 1.5;
-        }
-
-        @include wider-than($phone-max) {
-            @include set-font(h2);
-
-            line-height: 1.5;
-        }
-    }
-
-    .question-number {
-        justify-self: right;
-    }
-
-    .quiz-card {
-        overflow: hidden;
-        padding: 0;
-        position: relative;
-
-        .content-row {
-            display: grid;
-            grid-auto-columns: 100%;
-            grid-auto-flow: column;
-            min-height: 72rem;
-            overflow: hidden;
-        }
-
-        .card-content {
-            @include normal-padding();
-
-            &:not(.hidden) {
-                display: grid;
-                grid-gap: 5rem;
-            }
-
-            &.final {
-                @include title-font(9rem);
-
-                align-items: center;
-                justify-items: center;
-            }
-        }
-
-        .big-checkmark {
-            bottom: 0;
-            max-width: calc(95% - 10rem);
-            position: absolute;
-            right: 0;
-        }
-
-        .bottom-border {
-            width: 100.5%;
-        }
-    }
-
-    .answers {
-        display: grid;
-        grid-auto-rows: 1fr;
-        grid-gap: 2rem;
-
-        @include wider-than($phone-max) {
-            grid-template-columns: 1fr 1fr;
-        }
-
-        @include width-up-to($phone-max) {
-            margin-bottom: 10rem;
-        }
-
-        @include width-between($phone-max, $tablet-max) {
-            margin-bottom: calc(13.8rem - 10.2vw);
-        }
-    }
-
-    .answer {
-        cursor: pointer;
-        overflow: hidden;
-        padding: 0;
-
-        &.guessed {
-            color: $guessed-color;
-            pointer-events: none;
-            text-decoration: line-through;
-
-            .card-content {
-                border-color: $guessed-color;
-            }
-        }
-
-        &:not(.guessed) .card-content {
-            border-image: linear-gradient(to bottom, os-color(green), os-color(light-blue)) 1 stretch;
-        }
-
-        .card-content {
-            @include normal-padding();
-
-            align-items: center;
-            border: 0;
-            border-left: 0.6rem solid transparent;
-            display: grid;
-            justify-items: center;
-            text-align: center;
-
-            @include width-up-to($phone-max) {
-                @include set-font(body-large);
-
-                line-height: 1.5;
-            }
-
-            @include width-between($phone-max, $tablet-max) {
-                @include body-font(2rem);
-
-                line-height: 1.5;
-            }
-
-            @include wider-than($tablet-max) {
-                @include body-font(2.4rem);
-
-                line-height: 1.5;
-            }
-        }
-
-        .longest-answer {
-            opacity: 0;
-        }
-
-        .longest-answer,
-        .displayed-answer {
-            grid-column-start: 1;
-            grid-row-start: 1;
-        }
-    }
-
-    .lower-left {
-        bottom: 0;
-        position: absolute;
-        width: calc(6.29vw + 9.2rem);
-    }
-}
