--- src/app/pages/hero-journey/sections/thanks/thanks.html
@@ -1,7 +1,0 @@
-<template args="model">
-    <div class="text-block">
-        <h2>{model.heading}</h2>
-        <div class="description" data-html="description"></div>
-    </div>
-    <img src="{model.image.image}" alt="{model.image.altText}">
-</template>
