--- src/app/pages/hero-journey/sections/thanks/thanks.js
@@ -1,21 +1,0 @@
-import componentType, {insertHtmlMixin} from '~/helpers/controller/init-mixin';
-import {description as template} from './thanks.html';
-import css from './thanks.css';
-
-const spec = {
-    template,
-    view: {
-        classes: ['thanks', 'hidden'],
-        tag: 'section'
-    },
-    css
-};
-
-export default class extends componentType(spec, insertHtmlMixin) {
-
-    init(model) {
-        super.init();
-        this.model = model;
-    }
-
-}
