--- src/app/pages/hero-journey/sections/thanks/thanks.scss
@@ -1,34 +1,0 @@
-@import 'pattern-library/core/pattern-library/headers';
-
-.hero-journey .thanks {
-    align-items: center;
-    background-color: os-color(gold);
-    grid-column-gap: $normal-margin;
-    grid-row-gap: 5rem;
-    padding: 4rem $normal-margin 0;
-
-    @include wider-than($phone-max) {
-        grid-template-columns: 1fr 1fr;
-        padding-right: 0;
-    }
-
-    &:not(.hidden) {
-        display: grid;
-        min-height: calc(100vh - 5.5rem);
-    }
-
-    .text-block {
-        @include wider-than($media-content-max) {
-            margin-left: calc(50vw - #{$content-max / 2});
-        }
-    }
-
-    h2 {
-        @include set-font(h1);
-    }
-
-    img {
-        align-self: end;
-        max-width: 100%;
-    }
-}
