--- src/app/pages/hero-journey/sections/try/try.html
@@ -1,15 +1,0 @@
-<template args="model">
-    <div class="boxed">
-        <div class="headline-group">
-            <h2>{model.heading}</h2>
-            <div class="blurb">{model.description}</div>
-        </div>
-        <div class="button-and-image">
-            <div class="pulsing">
-                <div class="pulse"></div>
-                <a class="btn primary" href="{model.linkUrl}" target="_blank">{model.linkText}</a>
-            </div>
-            <img src="{model.image.image}" alt="{model.image.altText}">
-        </div>
-    </div>
-</template>
