+++ src/app/pages/press/article/article.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import BodyUnit from '~/components/body-units/body-units.jsx';
+import BodyUnit from '~/components/body-units/body-units';
 import {Byline} from '~/components/byline/byline';
 import {formatDateForBlog as formatDate} from '~/helpers/data';
 import '~/pages/blog/article/article.css';
