+++ src/app/pages/separatemap/search-box/inputs/inputs.js
@@ -99,58 +99,3 @@ export default function Inputs({
         </div>
     );
 }
-
-// export default class extends componentType(spec, busMixin) {
-//
-//     whenPropsUpdated(obj) {
-//         if ('textValue' in obj) {
-//             this.syncSearchInputs(this.textValue);
-//         }
-//         this.update();
-//     }
-//
-//     syncSearchInputs(value) {
-//         const els = Array.from(this.el.querySelectorAll('.search-input'));
-//
-//         els.forEach((el) => {
-//             if (el.value !== value) {
-//                 el.value = value;
-//             }
-//         });
-//     }
-//
-//     @on('input .search-input')
-//     handleInput(event) {
-//         this.emit('search-value', event.delegateTarget.value);
-//     }
-//
-//     @on('keypress .search-input')
-//     handleReturn(event) {
-//         if (event.key === 'Enter') {
-//             this.emit('run-search');
-//         }
-//     }
-//
-//     @on('click .search-clear')
-//     clearSearchInput() {
-//         this.emit('search-value', '');
-//         this.emit('run-search');
-//     }
-//
-//     @on('click .search-icon')
-//     toggleMinimize(event) {
-//         this.emit('toggle-minimized');
-//         event.delegateTarget.blur();
-//     }
-//
-//     @on('click .filter-toggle')
-//     toggleFilters(event) {
-//         this.emit('filter-toggle');
-//     }
-//
-//     @on('keypress [role="button"]')
-//     simulateClick({delegateTarget: target, key, preventDefault}) {
-//         $.treatSpaceOrEnterAsClick({target, key, preventDefault});
-//     }
-//
-// }
